png-reader/pyproject.toml
[tool.poetry]
name = "png-reader"
version = "0.1.0"
description = ""
authors = ["David Yue <davidyue5819@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.1.0"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
